Convert code to a pypip compatible module/project
This remains just a command line tool for now but it can be built and uploaded to pypi via hatch (https://hatch.pypa.io).

To test the module, a local copy can be installed with:
   pip install -e <folder where this repo is found>

It should be able to be built with:
   hatch build

and then published to pypi with:
  hatch publish

(this will require a pypi account under which the project will be published).

Once published, anyone can install the project with:
   pip install pymod

and use it via the command line simply with ‘pymod’.